requirements.txt
anthropic==0.51.0
mcp==1.8.0
python-dotenv==1.1.0
